% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyinit.R
\name{pyinit}
\alias{pyinit}
\title{PY (Pena-Yohai) initial estimates for S-estimates of regression}
\usage{
pyinit(X, y, intercept = TRUE, deltaesc = 0.5, cc.scale = 1.54764, prosac,
  clean.method = c("threshold", "proportion"), C.res, prop, py.nit, en.tol,
  mscale.maxit = 200, mscale.tol = 1e-08, mscale.rho.fun = c("bisquare",
  "huber"))
}
\arguments{
\item{X}{the data matrix X.}

\item{y}{the response vector.}

\item{intercept}{should an intercept be included in the models. Defaults to \code{TRUE}.}

\item{deltaesc, cc.scale}{parameters for the M-equation of the scale. If \code{cc.scale}
is missing or invalid, it will be chosen such that the expected value of the
rho function under the normal model is equal to \code{deltaesc}. To specify the
rho function, see parameter \code{control}.}

\item{prosac}{the proportion of observations to remove based on PSCs.}

\item{clean.method}{how to clean the data based on large residuals.
If \code{"threshold"}, all observations with scaled residuals larger than
\code{C.res} will be removed (\code{C.res} corresponds to the constant
\eqn{C_1} from equation (21) in Pena & Yohai (1999).
If \code{"proportion"}, observations with the largest \code{prop} residuals
will be removed.}

\item{C.res, prop}{see parameter \code{clean.method} for details.}

\item{py.nit}{the maximum number of iterations to perform.}

\item{en.tol}{the relative tolerance for convergence.}

\item{mscale.maxit}{Maximum number of iterations allowed for the m-scale algorithm.}

\item{mscale.tol}{Convergence threshold for the m-scale}

\item{mscale.rho.fun}{The rho function to use for the m-scale.}
}
\value{
\item{initCoef}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for S-estimates of regression.
}
\references{
Pena, D., & Yohai, V.. (1999). A Fast Procedure for Outlier Diagnostics in Large
Regression Problems. \emph{Journal of the American Statistical Association}, 94(446),
434-445. \url{http://doi.org/10.2307/2670164}
}
